requirements.txt
streamlit
pypdf2
sentence-transformers
faiss-cpu
transformers
torch
python-dotenv
numpy